moved useClickOutside to utils
