+++ web/surfing/src/util/index.ts
@@ -4,7 +4,7 @@
 
 // export { default as Web3Provider, useWallet } from "./hooks/useWeb3State";
 // export { fluRelayEnvironment, useWinningTransactions } from './api';
-export { useViewport } from "./hooks";
+export { useViewport, useClickOutside } from "./hooks";
 export type { SupportedChainsList } from "./chainProviders";
 
 export { SupportedChains } from "./chainProviders";
